1.2.7 -- dropped objects stop flashing
Field: "anyone that is placing objects down and aren't the host, their
objects are spazzing and disappearing/reappearing" -- including props
nobody was touching, and clearing up by itself for one skater at random.

Measured: 15,056 spawns against 15,044 destroys in 50 minutes for about
nine live objects, peaking at 81 of each in one second; 78 full sets
arriving and 144 going out per second; 521,000 packets. Against that, a
completely clean wire (0 rejected, 0 partsDropped, parts in order), the
sender's enumerate stable at array=1 own=1 with every skip counter zero,
no slot churn, no map-away flips. Nothing was being lost or filtered --
the same correct set was applied over and over, and each application is
a mark-and-sweep that destroys and respawns.

Cause: the 1.2.3 hash heartbeat could not converge. SetHash covers
position and rotation, but in.set -- the copy it is compared against --
was written in exactly one place, the completion of a full set, while
kMove/kPlace/kGone changed the objects underneath it. So the first nudge
of any prop left the two ends permanently unequal. Independently,
buildSetPart drops a record whose name the wire will not carry, so the
assembled count could never equal the count the heartbeat quoted.

Then the rate: answering an ask went through the same forceResync flag
as the beat, so answering also pinged, which provoked the next ask on
the next tick. With the dropper open the enum period is 33 ms, so a
disagreement that should have cost one set every six seconds cost thirty
a second.

- The cached set tracks the deltas applied on top of it (setUpsert /
  setErase), with the sender's append-and-compact-in-order discipline,
  because the hash walks records in order.
- SetHash and the new SendableCount skip records the wire refuses, so
  both ends describe the same bytes and the same number of them.
- forceResync splits into beatDue / answerWants / changed. A full set
  goes to everyone only on a real change, a ping only on the beat, an
  ask is answered once and without pinging back. Load-bearing: a gate
  that can never agree now degrades to one set per peer per six seconds,
  which is the blind re-send it replaced. The unbounded loop was the
  defect; the disagreement was only its trigger.
- SetSendInProgress no longer forces a resync. Re-entering the publish
  re-measured the set, and SendSet restarts from part 0 when the hash
  moves, so a large set being dragged restarted every frame and could
  never finish.
- EnumerateOwn returns negative for "could not read" instead of 0. It
  returned 0 for that and for "owns nothing" alike, and the publisher
  diffs, so one unreadable frame broadcast a whole park as removed.

Diagnostics: the [drop] peer N set line deduped on a process-wide
last-value, so the peer index printed and the set described could be
different peers -- it made a peer whose count never changed look like it
was alternating. Per-peer now. A new rate-limited churn line says
whether a receiver is losing objects or being handed new identities for
ones it already holds; those look identical from a spawn counter and are
opposite bugs.

Gate: omp_looptest now covers move, removal and unsendable-record
convergence. The existing drop-sync check could not have caught this --
it only ever pinged a set nothing had touched. Negative-control
verified: backing the delta upkeep out fails with "after a move the held
set no longer hashes like the sender's".

Also silences five C5055 enum-to-float deprecations in the pause menu's
slider rows.

+++ src/ui/pause_menu.cpp
@@ -2652,10 +2652,10 @@ static bool handleValueChange(void* params, bool isSlider) {
         if (isSlider && (k == g_voiceRangeKey || k == g_voiceVolKey || k == g_voiceSensKey)) {
             const float pct = *(const float*)((const uint8_t*)params + off::kChangeParamsNew);
             if (k == g_voiceRangeKey && g_voiceRangeKey) {
-                const float v = MPVOICE_RANGE_MIN + pct * (float)(MPVOICE_RANGE_MAX - MPVOICE_RANGE_MIN);
+                const float v = (float)MPVOICE_RANGE_MIN + pct * (float)(MPVOICE_RANGE_MAX - MPVOICE_RANGE_MIN);
                 MpPrefs_SetVoiceRangeM((int)(v + 0.5f));
             } else if (k == g_voiceVolKey && g_voiceVolKey) {
-                const float v = MPVOICE_VOL_MIN + pct * (float)(MPVOICE_VOL_MAX - MPVOICE_VOL_MIN);
+                const float v = (float)MPVOICE_VOL_MIN + pct * (float)(MPVOICE_VOL_MAX - MPVOICE_VOL_MIN);
                 MpPrefs_SetVoiceVolume((int)(v + 0.5f));
             } else if (g_voiceSensKey) {
                 MpPrefs_SetVoiceSensitivity((int)(pct * 100.0f + 0.5f));
@@ -2668,13 +2668,13 @@ static bool handleValueChange(void* params, bool isSlider) {
             // game prints.
             const float pct = *(const float*)((const uint8_t*)params + off::kChangeParamsNew);
             if (k == g_nameDistKey && g_nameDistKey) {
-                const float v = MPNAME_DIST_MIN + pct * (float)(MPNAME_DIST_MAX - MPNAME_DIST_MIN);
+                const float v = (float)MPNAME_DIST_MIN + pct * (float)(MPNAME_DIST_MAX - MPNAME_DIST_MIN);
                 MpPrefs_SetNameDistM((int)(v + 0.5f));
             } else if (k == g_bubbleDistKey && g_bubbleDistKey) {
-                const float v = MPBUBBLE_DIST_MIN + pct * (float)(MPBUBBLE_DIST_MAX - MPBUBBLE_DIST_MIN);
+                const float v = (float)MPBUBBLE_DIST_MIN + pct * (float)(MPBUBBLE_DIST_MAX - MPBUBBLE_DIST_MIN);
                 MpPrefs_SetBubbleDistM((int)(v + 0.5f));
             } else if (g_bubbleTextKey) {
-                const float v = MPBUBBLE_TEXT_MIN + pct * (float)(MPBUBBLE_TEXT_MAX - MPBUBBLE_TEXT_MIN);
+                const float v = (float)MPBUBBLE_TEXT_MIN + pct * (float)(MPBUBBLE_TEXT_MAX - MPBUBBLE_TEXT_MIN);
                 MpPrefs_SetBubbleTextSize((int)(v + 0.5f));
             }
             return true;
